feat(event): get event container
