+++ scripts/shell/coverage-mac.sh
@@ -1,0 +1,5 @@
+. scripts/shell/env-mac.sh
+
+npm run build:debug
+
+npm run test:coverage
