+++ test/event.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 /* eslint-disable no-undef */
-import { cancelBubble } from "../lib/index.esm";
+import { cancelBubble, getDefineListeners } from "../lib/index.esm";
 
 describe("cancelBubble", () => {
   it("should call stopPropagation if available", () => {
@@ -26,5 +26,32 @@ describe("cancelBubble", () => {
     cancelBubble(eventMock);
 
     expect(eventMock.cancelBubble).toBe(true);
+  });
+});
+
+describe("getDefineListeners", () => {
+  beforeEach(() => {
+    // Reset the global variable before each test
+    window.MAZEY_DEFINE_LISTENERS = undefined;
+  });
+
+  it("should return an empty object if defineListeners is not defined", () => {
+    const defineListeners = getDefineListeners();
+    expect(defineListeners).toEqual({});
+  });
+
+  it("should return the existing defineListeners object if it is defined", () => {
+    const existingListeners = { event1: () => {}, event2: () => {} };
+    window.MAZEY_DEFINE_LISTENERS = existingListeners;
+
+    const defineListeners = getDefineListeners();
+    expect(defineListeners).toBe(existingListeners);
+  });
+
+  it("should create a new defineListeners object if it is not an object", () => {
+    window.MAZEY_DEFINE_LISTENERS = "not an object";
+
+    const defineListeners = getDefineListeners();
+    expect(defineListeners).toEqual({});
   });
 });
